* Added some new utility functions to tcutil.h (hence the change to so many files.)
* Modified the translation of ++x for zero-terminated arrays so that
we avoid the overheads of the general case (e.g., x+e) which requires
a NULl-check and a loop.  In particular, ++x (and x++ in contexts
where the result is unused) gets mapped to something like:
"if (x != NULL && *x != 0) ++x; else _throw_arraybounds();".
The test for NULL is eliminated if the flow analysis can prove that
the array is not NULL.  I had to modify the flow analysis to attach
the necessary information to the expression.

As a result, stepping through a zero-terminated array is very fast.
For instance, in tests/shootout/cyc-simple_hash_cyc.h I modified the
string operations (e.g., mystrcmp) to take thin, not-null, @zeroterm
pointers and got a substantial speedup over the fat-pointer version.
In part, this is because the cost of calculating the numelts(-) of the
strings is so expensive and the strings typically being compared are
short.

This suggests that @zeroterm might actually be useful in other
places for performance reasons.  I'm going to try a version of
lib/string.cyc that uses them as much as possible to see what
effect this has.  In the meantime, maybe we should put our plans
to nuke @zeroterm on hold...

All of the tests ran except for two failures in tests/zeroterm.cyc.
The tests failed because we now throw Array_bounds instead of
Null_Exception.  I changed the test to reflect this and put in
a FIX comment.