Some improvements to how subscripting and dereferencing are treated for zeroterm pointers and arrays.  In a RHS-context, a zeroterm pointer with a known bound greater than 1 gets treated like a normal thin pointer, so there's no implicit calculation of the "true" end (i.e., where the zero is.) In turn, this allows us to eliminate null and bounds checks for (thin) zeroterm pointers.
For zeroterm pointers with bounds of 1, we calculate the
bound the slow way.

In a LHS-context, an update to a zeroterm pointer is still
very expensive.  At this point, all I can suggest is that
you coerce to a @nozeroterm pointer and wrip out the code
that tries to write the trailing 0.