Fixed bugs that Trevor found -- one involved typeof(exp).  The problem is that typeof(exp) may occur within a struct/union/typedef that is declared within a function, and the exp may refer to some variable local to the function.  We lift the type declaration out to the top-level in the toc phase, which could result in a free reference.  To avoid the problem, I simply replace the occurrence of typeof(exp) with exp's translated type.
Also, I added support for GCC's __extension__.  This is supposed
to surpress "pedantic" warnings for particlar declarations or
expressions.  I made it an expression form and also allow it
at the beginning of declarations.  However, only the expression
forms get propagated through right now.  That is, __extension__
on a declaration will be dropped in the resulting C code.

I did not change buildlib to keep it from filtering out __extension__
because I'm not sure these are the only places where the keyword
is allowed and I didn't want to break anything right now.