In Tcutil, fixed bugs in zeroable_type and eliminated supports_default per earlier email.
Changed this type definition, which simplified a number of gunky
things in the compiler:
  EXTERN_ABSYN struct Fndecl {
    scope_t    sc;         // static, extern, etc.
    bool       is_inline;  // inline flag
    qvar_t     name;       // function name
    stmt_t     body;   // body of function
    fn_info_t  i;
    // type differs from i in that it may have fewer attributes and it
    // may have a different const-ness on the return qualifier
    type_opt_t cached_type; // cached type of the function
    opt_t<list_t<vardecl_t>>   param_vardecls;// so we can use pointer equality
    struct Vardecl            *fn_vardecl; // used only for inner functions
  };
Perhaps the biggest negative is that function-arguments have types
that allow their names to be NULL even though they are never NULL.  On
the other hand, enforcing this statically previously made code-sharing
(for fndecls and fntypes) awkward.

Changed tests/Makefile to use CYCFLAGS for the failtests.  Feel free
to undo if that was the wrong thing to do.