Modified pattern checker to consider enum constants as if they were integers so that you get an inexhaustive match unless you have a default.  I realize this is less than ideal, but since we do not enforce that an enum value is actually one of the tags, this is necessary to ensure the soundness of the flow analysis.
I would suggest that we consider adding something like @strict enum's
which would allow you to define symbolic constants and avoid this
problem in the pattern matching, but prevent casts from other integral
types to strict enum types.  Then again, one could always just
use a datatype for this case (though it's a little more painful
to do this.)

I had to fix a number of places in the compiler and libraries (and
tests/streambuff.cyc) where we did pattern matching on enums to add a
default case.  I think that for maintenance purposes, we would be
better off using datatypes (unless we go back and add the @strict
annotation or something like that.)