Found the bug that was causing cyctest to coredump.  This one had nothing to do with attributes (which makes me think the calling-convention issue was something I was just (un)lucky enough to notice.)
This was a trivial thing: when I rewrote how we generate the type_t
globals in absyn.cyc, I got the mapping from ints to float-types wrong
(using 1, 2, 3 instead of 0, 1, 2).  So we gave things the wrong
type.  And that in turn made the injection for aprintf use the wrong
datatype field, which led to a core dump.