An attempt to figure out and fix matrix multiply.
Changed the way we promote the a comprehension when the bound is
of tag_t so that we avoid an unnecessary cast to unsigned int.
(That wasn't the problem.)

Changed the way we print the ThinCon type constructor (so that
I could debug things.)  In particular, it now prints the number
of elements.  Changed comparison of expressions to skip over
a cast to unsigned int for a valueof(-) expression.  This wasn't
the problem either.

Changed matrix multiply.  It had a type of the form:
  int@{cols}@{rows}
within the body of the function, where cols and rows
were variables of tag_t type.  Unfortunately, we don't
really support this within functions, only within prototypes
(as a hack from the parser.)  Sigh.

Got rid of warning in the port code.  There appears to be a lot
more of these in toc.cyc (probably Dan's merge of fn's and
fn types since variables become variable options.)

Mike, it would help if you could augment the makefile for
the shootout (and in cyclone_internal) so that I could easily
build a single target...  Not sure what the best way to do
this is.