Various updates and changes to the shootout benchmarks:
1) Changed cyclone/ary3.cyc to use dependent types.  However, it has
two bugs that need to be fixed.  One is that we can't actually
free the allocated data because int @{valueof(`n)} is not <= int @.
The only reason I could see that this should be not allowed is if
valueof(`n) could be zero, which it cannot in this case.  There is
also a problem that the relations apparently can't tell that
n-1 < n for an array of size n (where n > 0), and so indexing n-1 is
OK.

2) Now generate nocheck versions of the Cyclone benchmarks for
comparison.

3) Adjusted the start and end timers for Java to not count argument
processing, as is the case with Cyclone.

4) Only do -nogc for cyclone/ and not cyclone-port/ (still need to
fix all the cyclone/ benchmarks to not use the GC to make this fair).

5) Fixed bugs in other Cyclone benchmarks to get them to work, e.g.
spellcheck and echo.  Main problem was off-by-one error for
zero-terminated buffers.

Right now, I've no idea why Cyclone's ackermann is twice as slow
as gcc's.  Must be the cache. :) I also don't have a good reason
as to why Cyclone's exception handling more than an order of magnitude
slower than gcc's; perhaps the push/pop handling and comparison
to tags really does add up.  Of course, Java is 4x slower than Cyclone.

Also, realloc() seems to really hurt us on some of these.  Is there
any reason we should have a native realloc built in to the runtime
system made available through the compiler?  This would avoid
some redundant zeroing of bytes, etc.