Added regexmatch shootout benchmark, and fixed wordfreq.  Attempted to get methcall to work, but the compiler just doesn't support the width subtyping we would need at the moment.  This prevents us from implementing objinst as well.  Since we have pthread support now, we could do prodcons, but I haven't bothered.  These are the only three not implemented now.
To get the java regexmatch to work, you need to download a regular
expression library; see the README for details.  In porting regexmatch
to Cyclone, I found a bug in the C version: it was setting the size of
a buffer to be too high.  Hee hee.  It also isn't freeing all of its
data (but neither is Cyclone at the moment).

We should think about memory management issues in general.  The C code
is being "charged" for freeing stuff at the end, but the GC'ed
programs may not actually GC.  Probably forcing a GC isn't fair
either.  Any ideas on how to resolve this dilemna?  In any case, we
should make the cyclone/ versions (if not cyclone-port/) do manual
memory management where possible.  Also, any bright ideas as to how
to best measure memory footprint?