Fixed bug in nocheck implemenation in cyc_include.h (wasn't doing a null check even when -DNO_NULL_CHECK wasn't on).
Changed pthread support so that is turned on or off at link-time
rather than configure-time.  If you link with -lpthread you'll getthe
thread-safe runtime, otherwise you'll get the default.  Still need to
configure with --with-pthreads. (Any reason to keep doing this?)

Updated documentation on extern "C include".  Added more support
so that you can define a C function in the main body and then
give it a Cyclone type in the cyclone_override part.  This should
allow us to eliminate the weird typedefs we now use for Core::mkfat
and friends (though I just now realized this and didn't do it).