* Fixed bugs in tests/patterns.cyc -- needed consume attribute on functions.  Added a main for this and added it to the setof things compiled and run by the Makefile.
* Put in an initializer for the array of pthread_t's in testexn,
since on Cygwin, a pthread_t is not bits-only.  That test fails
on Cygwin, but perhaps because I'm running the wrong compiler or
something.

* Modified the pretty-printer and ugly-printer so that they put in
extra braces around declarations that introduce a variable that would
shadow the scope of an already declared local variable (only when
spitting out C code.)  Cleaned up the pretty-printer's treatment of
blocks a little bit.

* Modified the pattern match compiler so that it ultimately uses the
same local variable names for the pattern-bound identifiers.  This
involved copying stuff from the temps that are generated into the
locally-declared identifiers.  I confirmed that gcc actually
eliminates the extra copies when -O is on.  This enables a lot more
source debugging (thanks for the suggestions Nikhil!).  The only snag
is that you can't get at the pattern bound variables while stepping
through a "where" clause on a switch pattern, since they use the
temps.

These changes blow up the size of the output (and hence slow compile
times) but not by much.  For parse_tab.cyc, the diff was about .3
seconds out of 4.0 in compile time for no optimization, and about .6
seconds out of 6.1 with optimization.  The ugly-printed code has about
10,000 (out of 500,000) new characters.  All in all, it seems worth
the change.