New files, so run configure.  tests/testasm did not compile for me.
* Fixed bug in assertion for strncpy.  We were requiring that n <=
numelts(src), but what we're supposed to do is copy up to strlen(src)
or n characters, whichever is smaller.

* Added a lookup_other_opt function to the hashtable library that lets
you use as a key an object in some other region.  This is basically a
hack to avoid allocating when building hash-cons tables (used below).

* Better support for comparing constant expressions.  I need to
impose some total order on the expressions (so I can look them
up in hashtables or other data structures).  Right now, evexp
doesn't really support this, but it claims to.  I tried changing
it so that it did the "right thing" but this broke the compiler
in strange ways that I wasn't able to track down yet.  So,
the result is that for expressions which we can evaluate at
compile time to an unsigned integer, it uses the "old" code
and for all other expressions, it uses the new code.

* Added a new module, vcgen.cyc with interface vcgen.h.  This is
intended to ultimately replace the assertion stuff that's happening in
the flow analysis.  However, it does not yet scale well for at least
one file (cyclone.cyc) and it doesn't have the Pratt constraint solver
integrated, so I've turned it off by default.  Yanling and I will be
hacking on this to get improvements in the array bounds checking (and
perhaps other things that need to be tracked in a more path-sensitive
fashion.)

We still have a very limited assertion language which only tracks
equality and inequalities (though now we have a more expressive term
language), and now the assertions support disjunction (so we can be
path sensitive) and can reason about "escaping" values to a limited
degree.  Hash-consing is used to preserve sharing which is necessary
to keep the assertions from blowing up into a 2^n tree when we have
n-conditionals in a row (c.f., cyclone.cyc's main().)

We use a very simple widening strategy to generate loop invariants
which effectively preserves relations that weren't "killed" across the
loop, but not much more.  The invariant generator uses a very dumb,
syntactic theorem prover.

I pulled an evil hack to handle non-structured control-flow (e.g.,
goto's.)  Basically, you lose all information at a label (i.e., the
pre-condition is set to true) which ensures that we have an invariant,
though obviously not a very strong one :-) Still, this lets me do a
good job of finding invariants for structured code and makes the
analysis fairly "pretty" (read easy to maintain.)

In the near future, I expect to add an @assert(-) expression which you
will be able to use to put in your own loop-invariants and
assertions.  So, if you want precise flow information for goto's,
you'll have to put an @assert(-) on the labels.  Similarly, if our
dumb invariants aren't good enough, you will be able to put a stronger
one on a loop.  I also hope to make it possible to support third-party
theorem provers so that you can write something like @prove[foo](A)
where foo is a third-party prover.  We'll invoke the prover foo with
the currently calculated assertion A' for that program point and ask
it to construct a proof P that A' => A.  We'll then check the proof P.
We may also parameterize the prover by an acceptable signature so that
in essence, you're getting a proof of A'=>A subject to the validity of
that signature.