* Fixed bug in _npop_handler which caused it to not pop more than one thing.
* The argument parsing bug (where options -no-seq-c an -pg had to be
placed after -un-gcc for some reason) has mysteriously disappeared.
I suspect something had to do with the _npop_handler stuff since
the Arg::parse code has nested try blocks.

* Fixed bug in toc.cyc where we were calling _pop_region with an argument
when it takes none.

* Reworked evexp.cyc's support for comparing expressions.  We now have
exp_cmp(e1,e2) and uint_exp_cmp(e1,e2).  The former can be used to
compare expressions that are to be placed in a hash-table or set or
whatever.  In the long run, it should never fail, but right now, there
are a couple of cases missing.  It's only exercised in vcgen.cyc so it
shouldn't matter for now.

exp_cmp() identifies constant expressions that evaluate to the same
value, but it does not force the values to be unsigned ints.  The
function uint_exp_cmp(e1,e2) is similar except that it essentially
requires that e1 and e2 be coerceable to unsigned ints (and static
expressions.)  So, for instance, exp_cmp(NULL,0) != 0 whereas
uint_exp_cmp(NULL,0) == 0.

* Some tuning for vcgen.cyc.  In addition to hash-consing the
assertions, we use lazy substitution.  However, we do all of the
substitutions at once for a given assertion to avoid crawling over the
tree multiple times.  We also memoize whether or not the assertion is
reduced (i.e., has all of the substitutions pushed through.)  For
cyclone.cyc, this cuts compilation time with --vcgen in half.  In
addition, I'm avoiding using exception handlers for hash-table lookups
which also seems to help a lot.  Added lots of comments and made all
of the internal functions static.  We can probably tune things even
more by doing a better hash function for the assertion nodes.  Next
up, I plan to integrate the Pratt constraint solver.  Then we should
be able to use the information to eliminate array bounds checks.

* Commented out tests/testasm.cyc in the tests/Makefile since it doesn't
compile on cygwin.  Also, the testexn compiles but doesn't run correctly
on cygwin.