Re-run configure.
Added a soon-to-be pass that will remove unused labels from the
generated C code, purely for the purpose of prettier code and perhaps
faster gcc compiles.  The "pass" is currently a no-op, but is
integrated into cyclone.cyc.  Ben Hindman, a UW undergraduate, has
this pass partly written so we want to "plug it in" to the current
version.

Some other random clean-up, including making the lex table about 33%
smaller.