All genfiles changed because I changed cyc_include.h
Added first bit of support for autorelease pools for reference
counting.  This is not yet sound; I still need to add a new kind for
autoreleased pointers, the purpose of which is made clear below.
There is an example in tests/test_currgn.cyc if you want to look
there.

You can declare an autorelease pool similarly to a LIFO region:

{ autopool<`r>;
  ...
}

Within the scope of the declaration, `r is the name of the topmost
autorelease pool.  The function autorelease() attaches a
reference-counted pointer to the topmost pool.  For example, if y has
type int *`RC, and I do

x = autorelease(y);

Then x has type int *`r, where `r is the name of the topmost
autorelease pool.  There are no aliasing restrictions on `r, so I can
use it freely within the scope of the pool declaration. To be able to
generically name the topmost pool, we use the region "constant" `A,
similar to `C for LIFO regions.  Thus, autorelease has type

`a ?`A autorelease(`a ?`RC) __attribute__((consume(1)));

The semantics of autorelease are that when the pool is destroyed (goes
out of scope), all of the autoreleased pointers it contains will have
their reference-counts decremented.  Thus, if I autorelease a pointer
with refcount 1, then I am able to freely use the autoreleased alias
until the conclusion of the scope, at which point the pointer will be
freed.  If I want the pointer to last beyond the pool's lifetime, I
need to increment its count.  I can do this with the routine
inc_refptr, which has type:

 `a ?`RC inc_refptr(`a::TA ?`r ptr);

This takes an autoreleased pointer, increments the reference count,
and returns a tracked pointer to the same storage.  Both the original
and the new copy can be used together, but the new copy can outlive
the scope of the autopool.

That's the basic idea, but of course it's unsound because nothing
prevents me from passing an arbitrary pointer into inc_refptr.  To
solve this problem, we need to introduce a new flavor A, in addition
to U, T and _ (default), where A <= _.  Thus the lattice is:

     T
    / \
   U   _
       |
       A

The idea is that a AR-kinded pointer can be used wherever a normal
region (R-kinded) pointer can be used.  However, we will change the
type of autorelease and inc_refptr to return and accept AR-kinded
pointers, respectively.  I will do this soon, but I wanted to check in
to let Trevor (and/or others) play until I get to it.  I also don't do
anything with outlives and pools yet, but doing so should be
straightforward.