* Fixed the pragma's to work again.
* Fixed the treatment of casts in vcgen so that we drop unnecessary
  casts on terms.

* Fixed a bug in vcgen.  When the pre-condition was 'false' we were
  ignoring the statement and returning false immediately, which would
  normally be the right thing to do.  However, since we don't really
  treat goto properly, it's not here.  In particular, we reset the
  pre-condition to 'true' when we encounter a label.  So, we can't
  cut off, but rather must descend into the statement to look for
  any labels.

At this point, it seems that the vcgen is as accurate as the old
relations code, except when goto is involved.  In particular, it
eliminates the same checks that the relations code eliminates in
parse_tab.cyc except for two that come right after labels (where
we lose all information right now.)  On the other hand, the vcgen
actually picks up two checks that the relations didn't :-)

I also tried reducing the graphs by adding in equivalence classes
for terms, but it just slowed things down, so I threw this change
away.