Hacked conversion to C so we compile under gcc 3.4.
Specifically, we need to output C code like

    ((t)f)(...)

This is a function application where the function
is cast to some type t.  And gcc 3.4 does not like
our t any more.  So instead we output

    ({ (t)f; })(...)

and gcc 3.4 is happy again.

Also: gcc 3.4 refused to compile cf_flowinfo.cyc
because of a shadowed local variable, so I renamed
it.