* Added support for C99 array dependency notation for functions.  A function such as int f(int n, int a[n]) elaborates into int f(int n, int @a) @requires{n <= numelts(a)}.  This isn't quite working yet when we have a constant for the bound -- something's busted in the old relations code which I'm ripping out anyway.
* Added a new library sexp.  The library is intended for use in
serializing/deserializing data structures to files or strings --
effectively an interchange format.  Instead of using a datatype, I've
coded it in an OO-style, including support for visitors.  I plan to
use this to code up the interface between vcgen and external provers.
It might be useful to use this for other things (e.g., spitting out
Cyclone interfaces?)  Anyway, I have support for all of the built in
arithmetic types as well as strings, symbols, and vectors of objects
(called tuples) but you can pretty easily extend the interface with
new kinds of objects by defining your own class.  I'm tempted to
generalize visitors by using dictionaries instead of a record.  If
anyone has ideas on how to do this better, I'd like your input.

* Added tests/testsexp.cyc.  This tests printing and parsing for
sexps, as well as comparisons and the hash-consing.  There's also a
few interesting visitors defined to show how to use them.  Also, I had
to change md5.cyc to compile with the first change mentioned above.  I
also added in testasm, testsexp, and multimethod to the tests that are
run, and made HAVE_PTHREAD=no by default since cygwin doesn't have it.