* Fixed pthread bug reported by Prodromos
GC with pthreads requires using redirected versions of pthread
functions. pthread_create --> GC_pthread_create etc. Modified
cyc_include.h and cyclone.cyc to rename these functions in this
setting.

* Added rvmalloc

For device drivers that want to use vmalloc to grab large frame
buffers etc. This function is the same as rmalloc except it uses
vmalloc instead of kmalloc. The vmalloc'd pages are maintained along
with the standard region pages and are freed using vfree when the
region is cleaned up. For non-linux-kernel code, rvmalloc is the
identical to rmalloc.