* Big bug in vcgen -- was not calculating the right VC for the false-branch of an == or != test. The old version was essentially returning "false" for all of these!
* You can use @assert__(e) in place of @assert(e).  The former is
necessary when you #include <assert.h> to avoid problems with the
macro that it defines.

* We add 0 <= t to the constraint graph for all unsigned terms t that are
entered into the graph.  This is useful for proving that when i = 0,
then i <= numelts(s).

* Added type information for primop terms.  This lets us get more useful
information from the types into the theorem prover.

* Changed the add_node and related primitives to return a bool so that
when they add inconsistent information (which really should never happen)
we can check for it.

* Changed the way the @ensures clauses is checked so that it uses
fewer clauses (and fixed a bug.)