* Added lazy negation for assertions (i.e., an explicit Not(-).)  This gets eliminated along with Kill and Subst nodes when we call reduce, instead of as a separate pass.  It also allows us to short-circuit things like not(not(A)) and not(or(not(A),not(B))) etc.
* Reworked pratt_prover so that the graphs track difference
constraints on equivalence classes of terms (instead of the raw
terms.)  This was intended to help with the congruence closure, but
I'm not sure that it was a big win.  One problem is that I still end
up adding the 4 constraints (x u<= y, x s<= y, y u<= x, y s<= x) to
the graph because the terms x and y might be compound (e.g., t1+3) and
so simply unifying them doesn't capture all of the information.  (It
might if we normalized the terms first, but it was easier to just add
in the constraints.)

* Added some support for congruence closure.  Basically, when we enter
a new term into the graph, we try to find any equivalent term.  For
instance, if the graph contains numelts(x), x, and y as terms, and
x and y are in the same equivalence class (i.e., x == y), then when we
enter numelts(y) into the graph, we will unify numelts(x) and
numelts(y), so they will share all constraints.  This captures a lot,
but is far from complete.  In particular, we may not yet know x == y
and later add this fact.  Then we need to discover that we should
unify numelts(x) and numelts(y) and more generally, f(...x...) and
f(...y...) when the ...'s are also equivalent.

I've added a naive procedure to do this, but have not yet invoked it
because it could be expensive (essentially cubic).  A better way might
involve back-pointers for sub-terms so that we can find all
of the terms in the graph that they are in.  It's just painful to
implement this and I suspect there's a better way to organize things.

Additionally, we may discover through the constraints that x <= y and
y >= x hold, so we should unify x and y (which in turn may cause
numelts(x) and numelts(y) to unify, etc.)  Right now, I'm not doing
this either but we'll need to add support for this sooner or later.
If we have good support for it, then it's not clear that the equivalence
classes are doing much good.

In short, the constraint solver is smarter about equality, but the
approach is pretty ad hoc.  I need to re-think how to do this properly
with Yanling.

* These changes allowed me to get rid of some more dynamic-assert's in
lib/string.cyc in favor of static assert's.  I think we'll be able to
get rid of almost all of the bounds checks in the string library sometime
soon.