Simpler and smarter constraint solver:
* Added a new function add_difference_const(g, t1, t2, c) which adds
to the graph t1 - t2 == c.  Care is taken to ensure that c actually
fits within an int when adding the constraints to the graph.

* Got rid of break_term in pratt_prover.  Now when we add a term of
the form t+c to the graph, we use add_difference_cost to add in the
fact that t - zero == c.  This greatly simplifies the add_constraint
code, as the propagation of constants (and the overflow) is handled in
the all-paths closure now.  We may be able to simplify add_constraint
further.

* Modified add_eq to try to simplify the equation.  For instance, when
we have t1+c == t2 we can call add_difference_const(g, t1, t2, -c)
which effectively solves the equation for t1.  The reason we do this
here instead of relying upon the inequations is that we can more
freely move constants from one side of an equation to another than
we can for an inequation.

* Modified congruence closure to account for constants (duh.)  The
problem is that when processing +(x,y), we were running through the
domain of the graph to see if we could find a +(z,w) such that x == z
and y == w.  But we weren't discovering that when x == 3, we could
equate +(x,y) to +(3,y) or better, when x == 3 && y == 2, we could
equate +(x,y) and 5.  That's because the constants don't occur in
the domain of the graph -- they're encoded as differences with the
distinguished zero node.  So the function find_eq_const was added
to make it possible to determine if a term is equal to a constant.
When it is, we add in the appropriate congruences.