* Added lots of description and the copyright to the beginning of Pratt Prover.
* Modified the functions to throw an exception instead of return a bool
when a graph is found inconsistent.  This is more robust so that we don't
forget to check the return code, if a little more expensive due to the
try/catch's necessary.  It also cuts out a lot of "if (!...) return false"
tests.

* Factored out the code that traverses a list of graphs, applies a function
to each one, catches the Inconsistent exception, and deletes the graph.