It would be good if we could clean up error messages with respect to things like @aquals and regions.  In particular, it would be nice if we had a pretty-printing mode that didn't print this detail when it wasn't relevant to an error message.
* Patched bug that prevented structarray.cyc from compiling.  When
compressing a valueof_t(...), we now get rid of unnecessary casts
(unsigned int)valueof(...) which allows valueof_t<(unsigned int)valueof(t)>
to reduce to valueof_t<t>.  This also seemed to fix the problems
with streambuff, so I added both of these back in to the tests/Makefile.
However, the require.cyc test is now broken -- something strange
is happening with kind checking and I haven't had a chance to
track down the problem.  Since @requires clauses on unions haven't
been integrated with the new vcgen stuff, I'm not too worried about
it right now, as I expect this stuff to change.  (Nikhil, you might
want to look at this to see if you can figure out if your recent
changes had an effect.)

* Fixed bug in prover where it considered numelts of a T*{n} to
always evaluate to n.  This is only true when the value
is non-null.  We could improve the analysis quite a bit by adding in
((v != 0) => numelts(v) == n) && (v == 0) => numelts(v) == 0).  But this
is a disjunction which is awkward to support at the point where we insert
type-derived information into a constraint graph.

* Changed vcgen so that when generating a bounds-check on a thin
pointer, we use the type information instead of numelts(-).  That
avoids the problem above, where numelts(-) isn't really defined on
a null, thin pointer as far as the prover is concerned.

* Added support for eliminating null-checks using vcgen.  This manages
to eliminate a lot more null-checks (170 total in the files compiled
by a make cyclone_src), most of which appear to involve redundant
reads to memory (e.g., x->y->z and x->y->w) which is not thread-safe,
but since the bounds check stuff isn't thread-safe either, it's worth
doing for now.  (The long term goal is to teach vcgen about
thread-local vs. shared memory.)  Interestingly, all 170 checks
were eliminated by the simple prover and did not require the
constraint solver.

* Specialization of 0 != t and t != 0 in prover when t is not signed
(i.e., an unsigned int or a pointer).  We replace this assertion with
0 U< t to avoid a disjunction.  (It would actually be sound to do this
always since when t != 0, 0 U< t regardless as to whether t is signed.
However, I'm not sure if this loses information.)  This change was
crucial for keeping the running time of insert_checks reasonable,
since the NULL-check stuff adds a lot of t != 0 constraints.  Without
it, whenever we go to prove any constraint, we end up doing a lot
of backtracking/copying.

We still need to modify vcgen and insert_checks to deal properly with
casts/coercions.  Right now, vcgen just considers that these things
might fail, but doesn't refine the context.  We also need to modify
vcgen to interpret the pattern matching.