* Added support for reaps
You will need to rerun ./configure.

Using a modified version of the bget (lib/bget.c) buffer allocator to
manage memory for reaps. For now, all regions are also reaps. This is
probably not optimal since there is some overhead associated with
using bget as the allocator -- for regions in which no memory is ever
freed bget just bumps a pointer as before. But it consumes two header
words for each chunk allocated. Plan on coming up with something to
control the allocation strategy for region i.e. reverting to the
simple allocation scheme for plain-vanilla regions.

  The changes are in the runtime system only. I've just
#define'd the region function names in cyc_include.h to the new reap
function names.

  Here are the some of the new types, taken from core.h

 void ufree(`a::TA *@aqual(UNIQUE) `H ptr) __attribute__((noliveunique(1)));
  /** [ufree] frees a heap unique pointer. */

 void rufree(region_t<`r> h, `a::TA *@aqual(UNIQUE) `r ptr) __attribute__((noliveunique(2)));
  /** [rufree] frees a reap unique pointer. */

`a ?@aqual(REFCNT) `r alias_refptr(`a::TA ?@aqual(REFCNT) `r ptr;{});
  /** [alias_refptr] bumps up refcnt on both heap and reap rc ptrs */

 void drop_refptr(`a::TA ?@aqual(REFCNT) `H ptr;{})
			 __attribute((noliveunique(1)));
  /** [drop_refptr] decrements refcnt for heap refcnt */

 void rdrop_refptr(region_t<`r> h, `a::TA ?@aqual(REFCNT) `r ptr;{})
				__attribute__((noliveunique(2)));
  /** [drop_refptr] decrements refcnt for reap refcnt */