Got rid of TupleType, Tuple_e, and Tuple_p internally.  These are translated to the anonymous struct counterparts as part of parsing, but we maintain a flag that lets us print things out as tuples for error messages and so forth.
For the most part, this cut a lot of code out of the compiler.  For
instance, we no longer have to translate tuples to structs in the
back-end, don't have to deal with them during the analyses, etc.
Unfortunately, this all turned out to be trickier than it looks
because we didn't support pattern matching on anonymous structs.  So a
fair amount of new code had to be written to deal with this.

But as a side-effect you can now use x.fi and x->fi to access the
components of a tuple instead of x[i] and the very awkward (*x)[i].
Similarly, you should be able to use patterns of the form {.fi = pi,
.fj = pj, ...}.  The subscript notation gets mapped to .fi during
type-checking which means that we don't have to worry about it
down-stream.  I would like to make the old subscript notation go away
entirely, so in the near future, I'm likely to introduce a warning if
you use it.

This has been tested on the compiler and tests/ except for
tests/testbget since I didn't have sources for this.

Next up: I'd like to change the treatment of datatype constructors in
a similar fashion.  That is, a constructor like Foo(bar_t,baz_t) would
really be shorthand for Foo{bar_t f1; baz_t f2;}.  Additionally, I'd
like to make it possible to extract members, with .f0 being the value
of the tag.  The ultimate goal is to make it possible to eliminate
pattern matching before control-flow and vcgen analysis, again
simplifying these somewhat.