* Added a shorthand notation for aquals
  The aliasing component of a kind is now deprecated -- a warning is
issued if you use kinds like ::TB.

  Here's a mapping from the verbose form to the shorthand

  `a::B : RESTRICTED >= aquals(`a) --->  `a::B\T
  `a::B : UNIQUE >= aquals(`a)  --->  `a::U\T
  `a::B : REFCNT >= aquals(`a) --->  `a::B\RC
  `q::Q : RESTRICTED >= `q     ---> `q::Q\T

 Basically, you can use \T,\U,\A,\RC whereever RESTRICTED etc. is
expected. These tokens are also allowed to immediately follow a type
variable of A,B,M and Q kind. In this case it is interpreted as an
upper bound on the top-most alias qualifier of the type variable.

 The `U and `RC shorthands for unique and refcnt'd heap pointers
remains.

* Fixed a couple of bugs

 Existentially bound tvars `a should have aquals(`a) <= ALIASABLE by
default. This was missing earlier. Also, wasn't checking for alias
free paths properly in patterns. These should be fixed now.

* Added a missing #define to cyc_include.h

 Forgot to add a #define for the bget reaps. This was causing gcc to
emit an unnecessary warning on toc.c. This explains why all the
genfiles changed :(