* aprof now works with reaps
  region_used_bytes etc. should work again now.

  Note that I've added --novcgen to the aprof Makefile to work around
the performance issue.

* added a command line switch -noreaps

  Changed the region implementation so that the pointers to the
allocation functions are carried around with the region
handle. This adds a level of indirection when calling _region_malloc
etc. but allows us to switch the region allocator behavior. Providing
the -noreaps flag causes the original bump-pointer implementation to
be used. Of course, with -noreaps, functions like Core::rufree are
just no-ops. The default uses the bget allocator.

 This was implemented using a small change to the code generator that
now outputs _new_region(int, char*) for a region creation -- the int
indicates which allocator to use; and the char* is the region name as
before. For dynamic regions this strategy doesn't quite work, so a
a preprocessor hack in core.h does the trick.

 The one reason to use -noreaps might be the ~120 bytes overhead that
bget consumes for each region, and the two header words for (chunk
size, previous chunk size) for each allocation. Even so, aprof
indicates that without frees, the memory profile of reaps is nearly
identical to normal regions.

 Note these new signatures in core.h:

  struct NewDynamicRegion<UNIQUE> _new_ukey(unsigned int noreaps_flag,
					    const char *file,
					    const char *func,
					    int lineno);

  struct NewDynamicRegion<REFCNT> _new_rckey(unsigned int noreaps_flag,
					     const char *file,
					     const char *func,
					     int lineno);

 If you use the new_ukey, new_rckey macros then this shouldn't make
any difference. But, if you call these functions directly, as some
externC code in the scheme interpreter does, then the new argument
does matter.