Okay, now aprof.cyc should compile in a reasonable amount of time (about 3 seconds on my Cygwin laptop.)  Turns out that we had a lot of problems to fix which were mostly my fault:
* Fixed bug in reduce for assertions.  Wasn't getting the substituted
terms in a negation right.

* Fixed performance bug in term hashing.  We were mapping all
non-integer constants to the same hash value.  This ended up putting a
lot of things in one bucket which triggered some problems with the old
hashtable code (which was inherited from O'Caml and *still* has these
problems!)

* Fixed performance bug in hashtables.  The old code was expensive
when bucket lengths got long (see above) because it was using
List::length(-) on each insertion to see if the buckets were too long.
The new version keeps the bucket list length with the bucket.  While I
was at it, I fixed a bunch of other things.  First, I flattened the
bucket list to avoid an extra allocation (which makes up for the extra
word per bucket that I added.)  Second, I made the insertion code for
resizing use a loop instead of recursion so we don't blow the stack.
Third, when resizing, I calculate the new maximum bucket length and
set the new trigger value to two times this.  That keeps us from
resizing over-and-over again if we happen to have a bad hash function
that maps a lot of values to the same bucket.

* Fixed performance bug in insert_checks/vcgen: we were clearing out
the hash-cons tables after running vcgen, but we want to keep that
information around until after insert_checks runs.

* Fixed performance bug in reduction for assertions.  We now memoize
the reductions to make sure we walk the dag instead of the tree.
I avoided doing this earlier because you have to memoize based on
three arguments, but it turns out to be important.  This is what
finally shaved off all of the remaining overhead in compiling aprof.