* Updated manual to reflect aquals changes
  Moved the unique pointer section out of the regions section and into
a new section called "Pointers with restricted aliasing". There's
probably still some inconsistencies in the manual with references to
the old kinds, `U, `RC regions etc. Will need to refine further.

* Fixed a couple of bugs in the aquals implementation

The parser was not creating the right structure for aqual_t<`q>. Found
a bug in tctyp.cyc for handling the shorthand notation for qualifier
bounds ... \U, \T etc.